refactor(management): remove legacy route stack
