--- packages/management/src/dtos/index.ts
@@ -1,26 +1,0 @@
-export {
-	BulkActionResultSchema,
-	DeleteJobSchema,
-	ErrorSchema,
-	toBulkActionResultDto,
-	toDeleteJobDto,
-} from './actions.js';
-export { CapabilitiesSchema } from './capabilities.js';
-export {
-	JobCursorPageSchema,
-	JobSchema,
-	JobSelectorSchema,
-	RescheduleJobRequestSchema,
-	toJobCursorPageDto,
-	toJobDto,
-} from './job.js';
-export {
-	QueueStatsSchema,
-	QueueViewSummaryListSchema,
-	toQueueStatsDto,
-	toQueueViewSummaryListDto,
-} from './queue-view.js';
-export {
-	SchedulerHealthSchema,
-	toSchedulerHealthDto,
-} from './scheduler-health.js';
