+++ packages/management/src/mappers/job.ts
@@ -1,0 +1,59 @@
+import type { CursorPage, PersistedJob } from '@monque/core';
+
+import type { JobCursorPageDto, JobDto } from '../schemas/index.js';
+import type { ManagementOptions } from '../surface/types.js';
+
+export async function toJobCursorPageDto<TContext>(
+	options: ManagementOptions<TContext>,
+	page: CursorPage,
+	context: TContext,
+): Promise<JobCursorPageDto> {
+	return {
+		jobs: await Promise.all(
+			page.jobs.map((job) => toJobDto(options, job as PersistedJob, context)),
+		),
+		cursor: page.cursor,
+		hasNextPage: page.hasNextPage,
+		hasPreviousPage: page.hasPreviousPage,
+	};
+}
+
+export async function toJobDto<TContext>(
+	options: ManagementOptions<TContext>,
+	job: PersistedJob,
+	context: TContext,
+): Promise<JobDto> {
+	const serializePayload =
+		options.serializePayloadByJobName?.[job.name] ?? options.serializePayload;
+	const payload = serializePayload
+		? await serializePayload({ job, payload: job.data, context })
+		: job.data;
+	const dto: JobDto = {
+		id: job._id.toHexString(),
+		name: job.name,
+		status: job.status,
+		payload,
+		nextRunAt: job.nextRunAt.toISOString(),
+		lockedAt: job.lockedAt ? job.lockedAt.toISOString() : null,
+		claimedBy: job.claimedBy ?? null,
+		lastHeartbeat: job.lastHeartbeat ? job.lastHeartbeat.toISOString() : null,
+		failCount: job.failCount,
+		failureReason: job.failReason ?? null,
+		createdAt: job.createdAt.toISOString(),
+		updatedAt: job.updatedAt.toISOString(),
+	};
+
+	if (job.heartbeatInterval !== undefined) {
+		dto.heartbeatInterval = job.heartbeatInterval;
+	}
+
+	if (job.repeatInterval !== undefined) {
+		dto.repeatInterval = job.repeatInterval;
+	}
+
+	if (job.uniqueKey !== undefined) {
+		dto.uniqueKey = job.uniqueKey;
+	}
+
+	return dto;
+}
