+++ packages/management/src/mappers/scheduler-health.ts
@@ -1,0 +1,10 @@
+import type { SchedulerHealthDto } from '../schemas/index.js';
+
+export function toSchedulerHealthDto(healthy: boolean): SchedulerHealthDto {
+	return {
+		status: healthy ? 'ok' : 'unavailable',
+		scheduler: {
+			healthy,
+		},
+	};
+}
