--- packages/management/src/openapi/index.ts
@@ -1,1 +1,0 @@
-export { getManagementOpenApiDocument } from './openapi.js';
