--- packages/management/src/request/index.ts
@@ -1,1 +1,0 @@
-export { normalizeManagementRequest } from './normalizer.js';
