+++ packages/management/src/surface/index.ts
@@ -16,17 +16,11 @@ export type {
 	ManagementAuthorizationInput,
 	ManagementAuthorize,
 	ManagementErrorDto,
-	ManagementHttpMethod,
 	ManagementMonque,
 	ManagementOpenApiContext,
 	ManagementOptions,
 	ManagementPayloadSerializationInput,
 	ManagementPayloadSerializer,
-	ManagementQueryValue,
-	ManagementRequest,
-	ManagementResponse,
-	ManagementRoute,
-	ManagementRouteOperation,
 	ManagementSurface,
 	QueueStatsDto,
 	QueueViewSummaryDto,
